Don't cache package resolutions in the lens picker.
We were caching package resolutions in the lens picker to make up
for slowness in package manager and watching for the package change
broadcast to invalidate that cache.

The problem with this approach is that if a package gets updated as
part of the update of the system image, the lens picker will never
get the broadcast and the cache is now invalid, which canlead to
strange situations. Realistically, the only way to avoid this is
to not cache the resolution results.

Test: Switched around apps, media and otherwise.
Change-Id: I21b717f5f48ce66e71fcdd1a962b574294fdb2b2